zeals_take_home_project/requirements.txt
google-cloud-bigquery==2.34.4
google-cloud-storage==2.2.1
pyarrow==9.0.0
pandas>=1.0